% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{AnalyzeContestSensitivity}
\alias{AnalyzeContestSensitivity}
\title{Analyze Contest Sensitivity}
\usage{
AnalyzeContestSensitivity(
  formula,
  data,
  respondent_id,
  contest_no,
  var_interest,
  target,
  weights = NULL
)
}
\arguments{
\item{formula}{An object of class "formula" representing the model to be fitted.}

\item{data}{A data frame containing the conjoint survey data.}

\item{respondent_id}{String name of the respondent ID column.}

\item{contest_no}{String name of the contest number column.}

\item{var_interest}{String name of the variable of analysis (the target AMCE to analyze the sensitivity).}

\item{target}{String, either "sign" or "significance".}

\item{weights}{Optional string name of the weights column.}
}
\value{
A list containing the following components:
\describe{
\item{ContestSensitivity}{A data frame containing the sensitivity analysis results, including original model estimates, the number of influential contests dropped to reverse the result, and the updated rerun estimates.}
\item{DroppedContests}{A vector containing the unique IDs of the specific contests that were dropped to achieve the reversal.}
\item{ProfileKey}{A data frame mapping the numeric \code{profile_id} to their original string \code{profile} combinations for reference.}
\item{RerunData}{A data frame containing the cleaned conjoint data utilized for the reruns, excluding the dropped metadata and influence scores.}
}
}
\description{
Evaluates data at the contest level to determine the sensitivity of AMCEs to the removal of influential contests.
}
