% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{AnalyzeRespondentSensitivity}
\alias{AnalyzeRespondentSensitivity}
\title{Analyze Respondent Sensitivity}
\usage{
AnalyzeRespondentSensitivity(
  fit.lm,
  segroup,
  var_interest,
  target = "sign",
  dropped_resp_list = FALSE
)
}
\arguments{
\item{fit.lm}{The linear model object estimated with lm() to be replicated.}

\item{segroup}{Vector of variable to group standard errors.}

\item{var_interest}{String name of the variable of analysis, (the target AMCE to analyze the sensitivity).}

\item{target}{Target string, either "sign" or "significance". Default is "sign".}

\item{dropped_resp_list}{Logical indicating whether to return the list of dropped respondents. Default is FALSE.}
}
\value{
If \code{dropped_resp_list=FALSE} (default), returns a data frame containing the sensitivity analysis results, including original estimates, the number of respondents dropped, and rerun estimates.
If \code{dropped_resp_list=TRUE}, returns a list containing:
\describe{
\item{out_df}{The sensitivity analysis results data frame.}
\item{DroppedRespondents}{A vector of the specific respondent IDs that were dropped.}
}
}
\description{
Evaluates data at the respondent level to determine the sensitivity of AMCEs to the removal of influential respondents.
}
