% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cdf_respondent_influences}
\alias{plot_cdf_respondent_influences}
\title{Plot CDF of Respondent Influences}
\usage{
plot_cdf_respondent_influences(
  fit.lm,
  segroup,
  var_interest,
  target = "sign",
  ndrop = NULL,
  main = NULL
)
}
\arguments{
\item{fit.lm}{The linear model object estimated with lm() to be replicated.}

\item{segroup}{Vector of variables to group standard errors.}

\item{var_interest}{String name of the variable of analysis (the target AMCE).}

\item{target}{String, either "sign" or "significance". Default is "sign".}

\item{ndrop}{Optional integer indicating the number of respondents dropped to be highlighted in red.}

\item{main}{Optional string for the main title of the plot.}
}
\value{
A data frame of respondents and their calculated influences, sorted in decreasing order.
}
\description{
Plots the cumulative distribution function (CDF) of respondent influences on either the sign or significance of the target AMCE. Highlights the dropped respondents and the median respondent.
}
